split stuff out of shell
