+++ frontend/src/components/map/Bag.tsx
@@ -1,4 +1,6 @@
+import { getTileHeight } from '@app/helpers/tile';
 import { UnityComponentProps, useUnityComponentManager } from '@app/hooks/use-unity-component-manager';
+import { WorldTileFragment, getCoords } from '@downstream/core';
 import { memo, useCallback, useMemo, useState } from 'react';
 
 export interface BagData {
@@ -40,5 +42,61 @@ export const Bag = memo(
         });
 
         return null;
+    }
+);
+
+export const Bags = memo(
+    ({
+        tiles,
+        selectedMobileUnitID,
+        selectedElementID,
+        onClickBag,
+    }: {
+        tiles?: WorldTileFragment[];
+        selectedMobileUnitID?: string;
+        selectedElementID?: string;
+        onClickBag: (id: string) => void;
+    }) => {
+        const bagComponents = useMemo(() => {
+            console.time('bagloop');
+            if (!tiles) {
+                return [];
+            }
+            const ts = tiles.map((t) => {
+                const coords = getCoords(t);
+                const rewardBags =
+                    (selectedMobileUnitID &&
+                        t.sessions.flatMap((cs) => {
+                            return cs.bags.filter((equipSlot) => {
+                                if (!cs.attackTile || cs.attackTile.tile.id !== t.id) {
+                                    return false;
+                                }
+                                // reward containing bags have an ID that is made up of 16bits of sessionID and 48bits of MobileUnitID
+                                // bagIDs are 64bits
+                                const mobileUnitIdMask = BigInt('0xFFFFFFFFFFFF'); // 48bit mask (6 bytes)
+                                const bagMobileUnitID = (BigInt(equipSlot.bag.id) >> BigInt(16)) & mobileUnitIdMask;
+                                const truncatedMobileUnitID = BigInt(selectedMobileUnitID) & mobileUnitIdMask;
+                                return bagMobileUnitID === truncatedMobileUnitID;
+                            });
+                        })) ||
+                    [];
+
+                return t.bagCount > 0 || rewardBags.length > 0 ? (
+                    <Bag
+                        id={`bag/${t.id}`}
+                        key={`bag/${t.id}`}
+                        height={getTileHeight(t)}
+                        corner={0}
+                        selected={selectedElementID === `bag/${t.id}` ? 'outline' : 'none'}
+                        onPointerClick={onClickBag}
+                        {...coords}
+                    />
+                ) : null;
+            });
+            console.timeEnd('bagloop');
+            return ts;
+        }, [tiles, selectedMobileUnitID, selectedElementID, onClickBag]);
+
+        return <>{bagComponents}</>;
     }
 );
