+++ frontend/src/components/map/Tile.tsx
@@ -1,5 +1,8 @@
+import { getTileHeight } from '@app/helpers/tile';
 import { UnityComponentProps, useUnityComponentManager } from '@app/hooks/use-unity-component-manager';
-import { memo, useMemo } from 'react';
+import { WorldTileFragment, getCoords } from '@downstream/core';
+import { memo, useCallback, useMemo, useState } from 'react';
+import { TileHighlight } from './TileHighlight';
 
 export interface TileData {
     q: number;
@@ -21,5 +24,87 @@ export const Tile = memo(
         });
 
         return null;
+    }
+);
+
+export const Tiles = memo(
+    ({
+        tiles,
+        selectedTiles,
+        onClickTile,
+    }: {
+        tiles?: WorldTileFragment[];
+        selectedTiles?: WorldTileFragment[];
+        onClickTile: (id: string) => void;
+    }) => {
+        const [hovered, setHovered] = useState<string | undefined>();
+        const hoveredTile = hovered && tiles ? tiles.find((t) => t.id === hovered) : undefined;
+
+        const enter = useCallback((id) => {
+            setHovered(id);
+        }, []);
+
+        const exit = useCallback((id) => {
+            setHovered((prev) => (prev == id ? undefined : prev));
+        }, []);
+
+        const tileComponents = useMemo(() => {
+            console.time('tileloop');
+            if (!tiles) {
+                return [];
+            }
+            const ts = tiles.map((t) => {
+                const coords = getCoords(t);
+                return (
+                    <Tile
+                        key={t.id}
+                        id={t.id}
+                        height={getTileHeight(t)}
+                        color="#7288A6"
+                        onPointerEnter={enter}
+                        onPointerExit={exit}
+                        onPointerClick={onClickTile}
+                        {...coords}
+                    />
+                );
+            });
+            console.timeEnd('tileloop');
+            return ts;
+        }, [tiles, enter, exit, onClickTile]);
+
+        return (
+            <>
+                {tileComponents}
+                {hoveredTile &&
+                    [hoveredTile].map((t) => {
+                        const coords = getCoords(t);
+                        return (
+                            <TileHighlight
+                                key={`hov-${t.id}`}
+                                id={`hov-${t.id}`}
+                                height={getTileHeight(t)}
+                                color="white"
+                                style="gradient_blue"
+                                animation="none"
+                                {...coords}
+                            />
+                        );
+                    })}
+                {(selectedTiles || []).map((t) => {
+                    const coords = getCoords(t);
+                    return (
+                        <TileHighlight
+                            key={`selected-${t.id}`}
+                            id={`selected-${t.id}`}
+                            height={getTileHeight(t)}
+                            color="white"
+                            style="gradient_outline"
+                            animation="none"
+                            {...coords}
+                        />
+                    );
+                })}
+            </>
+        );
     }
 );
