+++ frontend/src/components/map/TileGoo.tsx
@@ -1,4 +1,6 @@
+import { GOO_SMALL_THRESH, getGooColor, getGooSize, getTileHeight } from '@app/helpers/tile';
 import { UnityComponentProps, useUnityComponentManager } from '@app/hooks/use-unity-component-manager';
+import { WorldTileFragment, getCoords } from '@downstream/core';
 import { memo, useMemo } from 'react';
 
 export interface TileGooData {
@@ -21,4 +23,38 @@ export const TileGoo = memo(({ id, q, r, s, height, color, size }: UnityComponen
     });
 
     return null;
+});
+
+export const TileGoos = memo(({ tiles }: { tiles?: WorldTileFragment[] }) => {
+    const tileGooComponents = useMemo(() => {
+        console.time('tileGooloop');
+        if (!tiles) {
+            return [];
+        }
+
+        const gs = tiles
+            .filter((t) => {
+                t.atoms.sort((a, b) => b.weight - a.weight);
+                return t.atoms.length > 0 && t.atoms[0].weight >= GOO_SMALL_THRESH;
+            })
+            .map((t) => {
+                const coords = getCoords(t);
+
+                return (
+                    <TileGoo
+                        key={`tileGoo-${t.id}`}
+                        id={`tileGoo-${t.id}`}
+                        height={getTileHeight(t) + 0.01}
+                        color={getGooColor(t.atoms[0])}
+                        size={getGooSize(t.atoms[0])}
+                        {...coords}
+                    />
+                );
+            });
+
+        console.timeEnd('tileGooloop');
+        return gs;
+    }, [tiles]);
+
+    return <>{tileGooComponents}</>;
 });
